Designed New FIFO
Last fifo used log2 function and thus only worked for sizes power of 2.
Also, it did not allow data to be written when the FIFO was full even if
data was going to be read on the same cycle.  The new one has neither of
these problems.

Still needs testing.