Git Housekeeping.  N/M parameters.  iSLIP update.
• A couple of files were renamed for clarity

• Files using 'RADIX' as a parameter or using numbers representing the
number of outputs or inputs to initiate vectors were changed to use N to
represent the number of inputs, and M to represent the number of
outputs.  This will make it easier to understand what the vectors
represent at a glance.

• Began creating an ifdef structure for various designs.

• Added a new PPE, the most basic PPE design, to take a priority input
vector.  This is used when creating an iSLIP allocator.

• Changed round robin allocator to true iSLIP