(ENoC) Overhaul of network.  Scripts added.  Test Results.
Changed file structure so that the network test had its own folder.

Added network vsim tcl scripts

changed the output enable of the fifo.  The network was single cycle, so
the input enable to an input queue was dependent on the output enable of
the downstream input queue.  Within the fifo, the output enable is
dependent on the input enable.  This causes a combinational loop
throughout the whole network.  Temporary fix is to remove that
dependency.

Ouput enable of the virtual output queue was causing prblems when
connected as a whole network.  This turned out to be that the o_en was
being assigned incorrectly (double).

Added results.  Poor saturation performance due to improper FIFO
management.  i.e, the way I broke the combinational loop as explained
above.  Will fix tomorrow.