(ENoC) Fixed comb loop.  near_full.  Retested.  Updated Scripts.
The combinational loop described in the previous commit was fixed.  The
output enable of the router was comb dependent on the output enable of
the input FIFO, which, in turn, was dependent on its input enable.  This
input enable was comb derived from the o_en of the downstream router
input FIFO.  One big loop.  Fixed by changing the router output enable
to be dependent on the full signals of the FIFO.  Did the same for VOQ.

Added a near_full signal to the FIFO_packet_t so that this could be
done.

I believe this may be causing problems however, as the FIFO could sample
a signal, but the router inputting this signal will not realise and will
send again on the next cycle.  However, thinking about it, the o_en from
the downstream router might ensure that no valid signal is sent to the
downstream router, thus this would not be a problem.  Needs
investigating.

Ran tests and updated scripts to show results.  Results saved in
ENoC>test>Results>18_03_2014.  Not perfect, tester needs improving.