(ENoC) Fixed en comb loop.
Fixed the combinational loop arising from the the fact that the output
enable of the FIFO was combinatorially dependent on its input enable.
Now, define DECOUPLE_EN as required.

Noticed a minor error in the switch control declaration.  The N and M
were mixed up in a declaration, causing confusion!  Fixed.

Edits to test bench ensured that simulated input nodes do not raise a
valid if they have no enable.

Added new test functions, one to give the maximum packet latency
measured and another to give the frequency of packet latencies.

Found small errors with the FIFO.   Added some flags to debug, found
that there was a digit wrong in the calculation of the output empty
flag.  Fixed.

Simplified the VOQ as the more complex version would cause delta errors.
i.e the valid would be given, which would effect the enable, which would
effect the valid etc.  Will return to this at some point.